Added support for movement "bump stops"
It's possible to have a character move not only based on time, but now also
based on x, y position bump stops. For example, it's possible to get the
ladybug character to move right at speed 0.5 until it reaches a given x
position. Likewise for the y direction and positions.

The syntax leaves a little to be desired, but this is already a great success.
We embed the entity in the path object now, so we don't have to do awkward
operations anymore, and a single `pathfinder.update(dt)` can also move all
the associated entities.